+++ src/test/java/name/richardson/james/bukkit/utilities/persistence/database/support/EntityChild.java
@@ -1,7 +1,7 @@
 /*******************************************************************************
  Copyright (c) 2013 James Richardson.
 
- TestBeanParent.java is part of bukkit-utilities.
+ EntityChild.java is part of bukkit-utilities.
 
  BukkitUtilities is free software: you can redistribute it and/or modify it
  under the terms of the GNU General Public License as published by the Free
@@ -16,38 +16,18 @@
  BukkitUtilities. If not, see <http://www.gnu.org/licenses/>.
  ******************************************************************************/
 
-package name.richardson.james.bukkit.utilities.persistence.database;
+package name.richardson.james.bukkit.utilities.persistence.database.support;
 
 import javax.persistence.*;
-
 import java.util.List;
 
-import com.avaje.ebean.validation.NotNull;
+import name.richardson.james.bukkit.utilities.persistence.database.AbstractRecord;
 
 @Entity
-public class TestBeanParent {
+public class EntityChild extends AbstractRecord {
 
-	@OneToMany(targetEntity = TestBeanChild.class, fetch = FetchType.EAGER, cascade = {CascadeType.PERSIST})
-	private List<TestBeanChild> children;
-
-	@NotNull
-	@Id
-	private int id;
-
-	public List<TestBeanChild> getChildren() {
-		return children;
-	}
-
-	public void setChildren(List<TestBeanChild> children) {
-		this.children = children;
-	}
-
-	public int getId() {
-		return id;
-	}
-
-	public void setId(int id) {
-		this.id = id;
-	}
+	@ManyToOne(targetEntity = EntityParent.class, fetch = FetchType.EAGER, cascade = {CascadeType.PERSIST})
+	@PrimaryKeyJoinColumn(name = "parentId", referencedColumnName = "id")
+	private List<EntityChild> children;
 
 }