+++ CHANGELOG.md
@@ -1,5 +1,8 @@
-# Changelog for BukkitUtilities (v6.0.5-SNAPSHOT)
+# Changelog for BukkitUtilities (v6.0.6-SNAPSHOT)
 
+
+**v6.0.5**  
+2013-12-14 15:04:22 +0000    commit changelog (James Richardson)  
 2013-12-14 14:18:39 +0000    remove duplicate statement (James Richardson)  
 2013-12-14 14:16:12 +0000    attempt to fix IndexOutOfBounds exception on Windows (James Richardson)  
 2013-12-14 14:06:34 +0000    fix various bugs in plugin updater (James Richardson)